refactor: solve menu circular dependency issue
